backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Authentication
bcrypt==4.1.1
PyJWT==2.8.0

# Firebase
firebase-admin==6.2.0

# HTTP requests
httpx==0.25.1
requests==2.31.0

# Data validation
pydantic==2.5.0

# ML libraries (for backend integration)
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
